PWN-7623 - move all ui stuff to ClaimUiMapper.kt
