+++ app/src/main/java/org/p2p/wallet/bridge/claim/ui/model/ClaimScreenData.kt
@@ -1,0 +1,8 @@
+package org.p2p.wallet.bridge.claim.ui.model
+
+data class ClaimScreenData(
+    val title: String,
+    val tokenIconUrl: String?,
+    val tokenFormattedAmount: String,
+    val fiatFormattedAmount: String,
+)
